fix: GPU 호환성 버그 수정 및 엔진 아키텍처 문서 추가
렌더링 버그 수정:
- SurfaceError::Outdated 처리 추가 (검정 화면 문제 해결)
- TAA Output 텍스처 COPY_DST 플래그 추가
- MaterialEvalLighting 버퍼 초기화 추가
- @invariant 추가 (zprepass.wgsl, visibility.wgsl)

셰이더 호환성 수정:
- textureLoad offset 파라미터 제거 (WGSL 호환성)
- textureSampleLevel 사용으로 변경
- 다수의 셰이더 파일 수정 (gtao, dof, ssr, ddgi 등)

기타:
- magic_impact.effect.ron: SphereVolume → Sphere
- docs/ENGINE_ARCHITECTURE.md: 엔진 구조 문서 추가

+++ src/shaders/dof_composite.wgsl
@@ -95,7 +95,7 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
                     let neighbor_coc = textureLoad(coc_texture, neighbor_pixel, 0).r;
                     if (neighbor_coc < 0.0) {
                         // Negative CoC = near field
-                        let distance = length(vec2<f32>(dx, dy));
+                        let distance = length(vec2<f32>(f32(dx), f32(dy)));
                         let neighbor_blur = abs(neighbor_coc) * params.screen_size.y;
 
                         if (distance < neighbor_blur) {
